feat(contribution): mise à jour de la modale provenant de l'extérieur
