+++ packages/code-du-travail-frontend/src/modules/contributions/ContributionAgreement.tsx
@@ -15,7 +15,7 @@ import {
   isAgreementSupported,
 } from "./contributionUtils";
 import { isExternalArrival } from "./externalArrival";
-import { ContributionGenericAgreementSearch } from "./ContributionGenericAgreementSearch";
+import { ContributionPersonalizedAgreementSearch } from "./ContributionPersonalizedAgreementSearch";
 import { useContributionTracking } from "./tracking";
 import { useRouter } from "next/navigation";
 import {
@@ -78,11 +78,14 @@ export function ContributionAgreement({ contribution, genericInfos }: Props) {
   } = useContributionTracking();
 
   // Chemin réel de la page (arbre classique `{num}-{slug}` et arbre « congés »)
-  // pour les actions Matomo émises par le bloc de sélection.
-  const trackingActionName = buildContributionAgreementPath(genericSlug, {
+  // pour les actions Matomo émises par le bloc de sélection. Suffixe `/extern` :
+  // ce bloc n'apparaît que dans le parcours externe (arrivée sur une page CC
+  // depuis l'extérieur ou après « Réinitialiser »), à distinguer du parcours
+  // interne de la fiche générique (suffixe `/intern`).
+  const trackingActionName = `${buildContributionAgreementPath(genericSlug, {
     num: parseInt(contribution.idcc, 10),
     slug: contribution.ccnSlug,
-  });
+  })}/extern`;
 
   // Fiche « générique » hybride pour le bloc de sélection : il raisonne sur la
   // fiche générique frère (slug de navigation + classification des CC).
@@ -136,9 +139,10 @@ export function ContributionAgreement({ contribution, genericInfos }: Props) {
   return (
     <>
       {mode === "selection" ? (
-        <ContributionGenericAgreementSearch
+        <ContributionPersonalizedAgreementSearch
           contribution={genericLikeContribution}
           personalizeTitleRef={personalizeTitleRef}
+          agreementName={contribution.ccnShortTitle}
           selectedAgreement={selectedAgreement}
           trackingActionName={trackingActionName}
           currentIdcc={contribution.idcc}
