+++ packages/code-du-travail-frontend/src/modules/contributions/ContributionGeneric.tsx
@@ -22,7 +22,10 @@ type Props = {
 export function ContributionGeneric({ contribution }: Props) {
   const [hash, setHash] = useState("");
   const personalizeTitleRef = useRef<HTMLParagraphElement>(null);
-  const getTitle = () => `/contribution/${slug}`;
+  // Suffixe `/intern` : les events de la fiche générique relèvent du parcours
+  // interne (parcours classique), à distinguer du parcours externe des pages CC
+  // (suffixe `/extern`, cf. ContributionAgreement).
+  const getTitle = () => `/contribution/${slug}/intern`;
   const { slug, isNoCDT, relatedItems } = contribution;
 
   const [displayGeneric, setDisplayGeneric] = useState(false);
